Put `AppInputDecoration` in separate file.
