+++ lib/otp/views/configure_hotp_dialog.dart
@@ -26,6 +26,7 @@ import '../../app/state.dart';
 import '../../core/models.dart';
 import '../../core/state.dart';
 import '../../oath/models.dart';
+import '../../widgets/app_input_decoration.dart';
 import '../../widgets/app_text_field.dart';
 import '../../widgets/choice_filter_chip.dart';
 import '../../widgets/responsive_dialog.dart';
